[LXR] Creating main loop tokenizer and `advance` method
In order to make things easier to tokenize the file, I added a field to
`Lexer` that is mapped to the current position of the file and a current
char. It will avoid things like passing the file or an iterator to a
method in order to access the source code.